promote PR ID and line number to int64
Fixes #2443

PR ID and line number are were previously downcasted to int32 in a
number of places despite being mostly treated as an int64 type. While
the likelihood of either of these values exceeding a 32-bit MAX_INT are
minimal, this change makes them 64-bit to err on the side of caution.

+++ internal/engine/eval/homoglyphs/communication/reviewer.go
@@ -82,7 +82,7 @@ func (ra *GhReviewPrHandler) SubmitReview(ctx context.Context, reviewText string
 // Hydrate hydrates the handler with a pull request
 func (ra *GhReviewPrHandler) Hydrate(ctx context.Context, pr *pb.PullRequest) {
 	logger := zerolog.Ctx(ctx).With().
-		Int32("pull-number", pr.Number).
+		Int64("pull-number", pr.Number).
 		Str("repo-owner", pr.RepoOwner).
 		Str("repo-name", pr.RepoName).
 		Logger()
