+++ internal/engine/eval/vulncheck/review.go
@@ -185,7 +185,7 @@ func newReviewPrHandler(
 	}
 
 	logger := zerolog.Ctx(ctx).With().
-		Int32("pull-number", pr.Number).
+		Int64("pull-number", pr.Number).
 		Str("repo-owner", pr.RepoOwner).
 		Str("repo-name", pr.RepoName).
 		Logger()
@@ -560,7 +560,7 @@ func newSummaryPrHandler(
 	cli provifv1.GitHub,
 ) (prStatusHandler, error) {
 	logger := zerolog.Ctx(ctx).With().
-		Int32("pull-number", pr.Number).
+		Int64("pull-number", pr.Number).
 		Str("repo-owner", pr.RepoOwner).
 		Str("repo-name", pr.RepoName).
 		Logger()
